# docs/swagger.yaml
definitions:
...
  kvm.CloudRegisterRequest:
    properties:
      clientId:
        type: string
      cloudApi:
        type: string
      oidcGoogle:
        type: string
      token:
        type: string
    type: object
...
info:
  contact: {}
  description: JetKVM backend API documentation
  title: JetKVM API
  version: "1.0"
paths:
...
  /cloud/register:
    post:
      consumes:
      - application/json
      description: Register the device to the cloud using a temporary token and OIDC
      parameters:
      - description: Cloud register request
        in: body
        name: CloudRegisterRequest
        required: true
        schema:
          $ref: '#/definitions/kvm.CloudRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AuthToken: []
      summary: Register device to cloud
      tags:
      - cloud
...
securityDefinitions:
  AuthToken:
    in: cookie
    name: authToken
    type: apiKey
swagger: "2.0"